requirements.txt
python-dotenv
langchain
langchain-openai
langchain-groq
pinecone
PyPDF2
langchain_google_firestore
langchain_chroma
faiss-cpu
sentence-transformers
fitz
streamlit
duckduckgo-search
langchain-pinecone